refactor: move tests to Car
